+++ tuxemon/npc.py
@@ -3,22 +3,20 @@
 from __future__ import annotations
 
 import logging
-import os
 import uuid
 from collections.abc import Iterable, Mapping, Sequence
 from math import hypot
 from typing import TYPE_CHECKING, Any, Optional, TypedDict
 
-from tuxemon import prepare, surfanim
+from tuxemon import prepare
 from tuxemon.battle import Battle, decode_battle, encode_battle
 from tuxemon.boxes import ItemBoxes, MonsterBoxes
-from tuxemon.compat import Rect
-from tuxemon.db import Direction, ElementType, EntityFacing, SeenStatus, db
+from tuxemon.db import Direction, ElementType, SeenStatus, db
 from tuxemon.entity import Entity
-from tuxemon.graphics import load_and_scale
 from tuxemon.item.item import Item, decode_items, encode_items
 from tuxemon.locale import T
 from tuxemon.map import dirs2, dirs3, get_coords_ext, get_direction, proj
+from tuxemon.map_view import SpriteRenderer
 from tuxemon.math import Vector2
 from tuxemon.mission import Mission, decode_mission, encode_mission
 from tuxemon.monster import Monster, decode_monsters, encode_monsters
@@ -28,8 +26,6 @@ from tuxemon.technique.technique import Technique
 from tuxemon.tools import vector2_to_tile_pos
 
 if TYPE_CHECKING:
-    import pygame
-
     from tuxemon.economy import Economy
     from tuxemon.states.world.worldstate import WorldState
 
@@ -128,10 +124,11 @@ class NPC(Entity[NPCState]):
         # pathfinding and waypoint related
         self.pathfinding: Optional[tuple[int, int]] = None
         self.path: list[tuple[int, int]] = []
+        # Stores the final destination sent from a client
         self.final_move_dest = [
             0,
             0,
-        ]  # Stores the final destination sent from a client
+        ]
 
         # This is used to 'set back' when lost, and make movement robust.
         # If entity falls off of map due to a bug, it can be returned to this value.
@@ -151,24 +148,7 @@ class NPC(Entity[NPCState]):
         # Move direction allows other functions to move the npc in a controlled way.
         # To move the npc, change the value to one of four directions: left, right, up or down.
         # The npc will then move one tile in that direction until it is set to None.
-
-        # TODO: move sprites into renderer so class can be used headless
-        self.playerHeight = 0
-        self.playerWidth = 0
-        # Standing animation frames
-        self.standing: dict[str, pygame.surface.Surface] = {}
-        # Moving animation frames
-        self.sprite: dict[str, surfanim.SurfaceAnimation] = {}
-        self.surface_animations = surfanim.SurfaceAnimationCollection()
-        self.load_sprites()
-        self.rect = Rect(
-            (
-                self.tile_pos[0],
-                self.tile_pos[1],
-                self.playerWidth,
-                self.playerHeight,
-            )
-        )
+        self.sprite_renderer = SpriteRenderer(self)
 
     def get_state(self, session: Session) -> NPCState:
         """
@@ -215,7 +195,7 @@ class NPC(Entity[NPCState]):
             save_data: Data used to recreate the NPC.
 
         """
-        self.facing = save_data.get("facing", Direction.down)
+        self.facing = Direction(save_data.get("facing", "down"))
         self.game_variables = save_data["game_variables"]
         self.tuxepedia = save_data["tuxepedia"]
         self.contacts = save_data["contacts"]
@@ -241,62 +221,7 @@ class NPC(Entity[NPCState]):
         self.template.slug = _template["slug"]
         self.template.sprite_name = _template["sprite_name"]
         self.template.combat_front = _template["combat_front"]
-        self.load_sprites()
-
-    def load_sprites(self) -> None:
-        """Load sprite graphics."""
-        # TODO: refactor animations into renderer
-        # Get all of the player's standing animation images.
-        self.interactive_obj: bool = False
-        if self.template.slug == "interactive_obj":
-            self.interactive_obj = True
-
-        self.standing = {}
-        for standing_type in list(EntityFacing):
-            # if the template slug is interactive_obj, then it needs _front
-            if self.interactive_obj:
-                filename = f"{self.template.sprite_name}.png"
-                path = os.path.join("sprites_obj", filename)
-            else:
-                filename = (
-                    f"{self.template.sprite_name}_{standing_type.value}.png"
-                )
-                path = os.path.join("sprites", filename)
-            self.standing[standing_type] = load_and_scale(path)
-        # The player's sprite size in pixels
-        self.playerWidth, self.playerHeight = self.standing[
-            EntityFacing.front
-        ].get_size()
-
-        # avoid cutoff frames when steps don't line up with tile movement
-        n_frames = 3
-        frame_duration = (1000 / CONFIG.player_walkrate) / n_frames / 1000 * 2
-
-        # Load all of the player's sprite animations
-        anim_types = list(EntityFacing)
-        for anim_type in anim_types:
-            if not self.interactive_obj:
-                images: list[str] = []
-                anim_0 = f"sprites/{self.template.sprite_name}_{anim_type.value}_walk"
-                anim_1 = f"sprites/{self.template.sprite_name}_{anim_type.value}.png"
-                images.append(f"{anim_0}.{str(0).zfill(3)}.png")
-                images.append(anim_1)
-                images.append(f"{anim_0}.{str(1).zfill(3)}.png")
-                images.append(anim_1)
-
-                frames: list[tuple[pygame.surface.Surface, float]] = []
-                for image in images:
-                    surface = load_and_scale(image)
-                    frames.append((surface, frame_duration))
-
-                _surfanim = surfanim.SurfaceAnimation(frames, loop=True)
-                self.sprite[f"{anim_type.value}_walk"] = _surfanim
-
-        # Have the animation objects managed by a SurfaceAnimationCollection.
-        # With the SurfaceAnimationCollection, we can call play() and stop() on
-        # all the animation objects at the same time, so that way they'll
-        # always be in sync with each other.
-        self.surface_animations.add(self.sprite)
+        self.sprite_renderer.load_sprites()
 
     def pathfind(self, destination: tuple[int, int]) -> None:
         """
@@ -411,8 +336,8 @@ class NPC(Entity[NPCState]):
 
         """
         # update physics.  eventually move to another class
+        self.sprite_renderer.update(time_delta)
         self.update_physics(time_delta)
-        self.surface_animations.update(time_delta)
 
         if self.pathfinding and not self.path:
             # wants to pathfind, but there was no path last check
@@ -445,7 +370,7 @@ class NPC(Entity[NPCState]):
         # TODO: its not possible to move the entity with physics b/c this stops that
         if not self.path:
             self.cancel_movement()
-            self.surface_animations.stop()
+            self.sprite_renderer.surface_animations.stop()
 
     def move_one_tile(self, direction: Direction) -> None:
         """
@@ -519,7 +444,7 @@ class NPC(Entity[NPCState]):
             # it still occasionally happens though!
             # eventually, there will need to be a global clock for the game,
             # not based on wall time, to prevent visual glitches.
-            self.surface_animations.play()
+            self.sprite_renderer.surface_animations.play()
             self.path_origin = self.tile_pos
             self.velocity3 = moverate * dirs3[direction]
             self.remove_collision()
